Adds the solve_order option to cga and cga_iteration
This option allows to make use of the fact that the update for
one player can obtained cheaply from the 'move' of the opponent.
This commit doesn't change the previous behavior as it defaults to
'both'. The 'alternate' mode switches between computing the update
for x and cheaply computing the one for y and then the other way
around in the next iteration. The modes 'xy' and 'yx' stay fixed.